+++ src/main/java/insane96mcp/mobspropertiesrandomness/data/json/util/MPRWeightedPreset.java
@@ -27,6 +27,8 @@ public class MPRWeightedPreset implements IMPRObject, IWeightedRandom {
 
 	private transient int _weight;
 
+	private MPRModifiableValue chance;
+
 	public MPRConditions conditions;
 
 	@Override
@@ -51,24 +53,29 @@ public class MPRWeightedPreset implements IMPRObject, IWeightedRandom {
 
 		if (this.conditions != null)
 			this.conditions.validate();
+
+		if (this.chance != null)
+			this.chance.validate();
 	}
 
 	/**
-	 * Returns this MPRWeightedPreset with the weight calculated based off the modifiers, or null if the world whitelist doesn't match
+	 * Returns this MPRWeightedPreset with the weight calculated based off the modifiers, or null if the world whitelist / chance doesn't match
 	 */
 	@Nullable
-	public MPRWeightedPreset computeAndGet(LivingEntity entity, Level world) {
+	public MPRWeightedPreset computeAndGet(LivingEntity entity, Level level) {
 		if (this.conditions != null && !this.conditions.conditionsApply(entity))
 			return null;
+		if (this.chance != null && level.random.nextDouble() >= this.chance.getValue(entity, level))
+			return null;
 
-		this._weight = (int) this.modifiableWeight.getValue(entity, world);
+		this._weight = (int) this.modifiableWeight.getValue(entity, level);
 
 		return this;
 	}
 
 	@Override
 	public String toString() {
-		return String.format("WeightedPreset{id: %s, weight: %s, conditions: %s}", this.id, this.modifiableWeight, this.conditions);
+		return String.format("WeightedPreset{id: %s, weight: %s, conditions: %s, chance: %s}", this.id, this.modifiableWeight, this.conditions, this.chance);
 	}
 
 	@Override